Move all search configs into single file
